LayoutTests/webgl/2.0.0/conformance2/renderbuffers/multisampled-renderbuffer-initialization-expected.txt

This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance2/renderbuffers/multisampled-renderbuffer-initialization.html
[ 1: PASS ] internal buffers have been initialized to 0
[ 2: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 3: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 4: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 5: PASS ] user buffer has been cleared to green
[ 6: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 7: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 8: FAIL ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) should be 36053. Was 36054.
[ 9: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 10: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 11: PASS ] user buffer has been initialized to 0
[ 12: FAIL ] getError expected: NO_ERROR. Was INVALID_FRAMEBUFFER_OPERATION : should be no errors
[ 13: PASS ] internal buffers have been initialized to 0
[ 14: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 15: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 16: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 17: PASS ] user buffer has been cleared to green
[ 18: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 19: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 20: FAIL ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) should be 36053. Was 36054.
[ 21: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 22: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 23: PASS ] user buffer has been initialized to 0
[ 24: FAIL ] getError expected: NO_ERROR. Was INVALID_FRAMEBUFFER_OPERATION : should be no errors
[ 25: PASS ] internal buffers have been initialized to 0
[ 26: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 27: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 28: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 29: PASS ] user buffer has been cleared to green
[ 30: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 31: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 32: FAIL ] getError expected: NO_ERROR. Was INVALID_FRAMEBUFFER_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 33: FAIL ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) should be 36053. Was 36054.
[ 34: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : should be no errors
[ 35: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 36: PASS ] user buffer has been initialized to 0
[ 37: FAIL ] getError expected: NO_ERROR. Was INVALID_FRAMEBUFFER_OPERATION : should be no errors
[ 38: PASS ] internal buffers have been initialized to 0
[ 39: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 40: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 41: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 42: PASS ] user buffer has been cleared to green
[ 43: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 44: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 45: FAIL ] getError expected: NO_ERROR. Was INVALID_FRAMEBUFFER_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 46: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 47: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 48: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 49: PASS ] user buffer has been initialized to 0
[ 50: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 51: PASS ] internal buffers have been initialized to 0
[ 52: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 53: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 54: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 55: PASS ] user buffer has been cleared to green
[ 56: PASS ] getError was expected value: NO_ERROR : should be no error after renderbufferStorageMultisample(RGBA8).
[ 57: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 58: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 59: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 60: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 61: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 62: FAIL ] user buffer has been initialized to 0 at (0, 0) expected: 0,0,0,0 was 0,255,0,255
[ 63: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 64: PASS ] internal buffers have been initialized to 0
[ 65: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 66: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 67: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 68: PASS ] user buffer has been cleared to green
[ 69: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 70: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 71: FAIL ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) should be 36053. Was 36054.
[ 72: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 73: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 74: PASS ] user buffer has been initialized to 0
[ 75: FAIL ] getError expected: NO_ERROR. Was INVALID_FRAMEBUFFER_OPERATION : should be no errors
[ 76: PASS ] internal buffers have been initialized to 0
[ 77: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 78: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 79: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 80: PASS ] user buffer has been cleared to green
[ 81: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 82: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 83: FAIL ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) should be 36053. Was 36054.
[ 84: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 85: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 86: PASS ] user buffer has been initialized to 0
[ 87: FAIL ] getError expected: NO_ERROR. Was INVALID_FRAMEBUFFER_OPERATION : should be no errors
[ 88: PASS ] internal buffers have been initialized to 0
[ 89: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 90: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 91: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 92: PASS ] user buffer has been cleared to green
[ 93: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 94: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 95: FAIL ] getError expected: NO_ERROR. Was INVALID_FRAMEBUFFER_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 96: FAIL ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) should be 36053. Was 36054.
[ 97: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : should be no errors
[ 98: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 99: PASS ] user buffer has been initialized to 0
[ 100: FAIL ] getError expected: NO_ERROR. Was INVALID_FRAMEBUFFER_OPERATION : should be no errors
[ 101: PASS ] internal buffers have been initialized to 0
[ 102: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 103: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 104: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 105: PASS ] user buffer has been cleared to green
[ 106: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 107: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 108: FAIL ] getError expected: NO_ERROR. Was INVALID_FRAMEBUFFER_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 109: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 110: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 111: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 112: PASS ] user buffer has been initialized to 0
[ 113: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 114: PASS ] internal buffers have been initialized to 0
[ 115: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 116: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 117: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 118: PASS ] user buffer has been cleared to green
[ 119: PASS ] getError was expected value: NO_ERROR : should be no error after renderbufferStorageMultisample(RGBA8).
[ 120: PASS ] getError was expected value: NO_ERROR : should be no error after framebufferRenderbuffer.
[ 121: FAIL ] getError expected: NO_ERROR. Was INVALID_OPERATION : should be no error after renderbufferStorageMultisample(RGBA8).
[ 122: PASS ] gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
[ 123: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 124: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 125: FAIL ] user buffer has been initialized to 0 at (0, 0) expected: 0,0,0,0 was 0,255,0,255
[ 126: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 127: PASS ] clearColor is [0, 1, 0, 1]
[ 128: PASS ] getError was expected value: NO_ERROR : should be no errors
[ 129: PASS ] successfullyParsed is true
[ FAIL ] 30 failures reported
